+++ docs/quality/changelog.md
@@ -76,3 +76,15 @@
 - `2026-02-26 15:15:10` [feat/T6-static-analysis-tools] **Write**: `/Users/a11401/Documents/2026/Mider/tests/test_tools/test_proc_runner.py`
 - `2026-02-27 10:07:23` [feat/T6-static-analysis-tools] **Edit**: `/Users/a11401/Documents/2026/Mider/tests/test_tools/test_eslint_runner.py`
 - `2026-02-27 10:07:28` [feat/T6-static-analysis-tools] **Edit**: `/Users/a11401/Documents/2026/Mider/tests/test_tools/test_eslint_runner.py`
+- `2026-02-27 10:11:12` [feat/T6-static-analysis-tools] **Edit**: `/Users/a11401/Documents/2026/Mider/mider/tools/static_analysis/eslint_runner.py`
+- `2026-02-27 10:11:16` [feat/T6-static-analysis-tools] **Edit**: `/Users/a11401/Documents/2026/Mider/mider/tools/static_analysis/clang_tidy_runner.py`
+- `2026-02-27 10:11:20` [feat/T6-static-analysis-tools] **Edit**: `/Users/a11401/Documents/2026/Mider/mider/tools/static_analysis/proc_runner.py`
+- `2026-02-27 10:11:28` [feat/T6-static-analysis-tools] **Edit**: `/Users/a11401/Documents/2026/Mider/mider/tools/static_analysis/eslint_runner.py`
+- `2026-02-27 10:11:43` [feat/T6-static-analysis-tools] **Edit**: `/Users/a11401/Documents/2026/Mider/mider/tools/static_analysis/proc_runner.py`
+- `2026-02-27 10:11:55` [feat/T6-static-analysis-tools] **Edit**: `/Users/a11401/Documents/2026/Mider/mider/tools/static_analysis/eslint_runner.py`
+- `2026-02-27 10:12:02` [feat/T6-static-analysis-tools] **Edit**: `/Users/a11401/Documents/2026/Mider/mider/tools/static_analysis/clang_tidy_runner.py`
+- `2026-02-27 10:12:06` [feat/T6-static-analysis-tools] **Edit**: `/Users/a11401/Documents/2026/Mider/mider/tools/static_analysis/proc_runner.py`
+- `2026-02-27 10:12:10` [feat/T6-static-analysis-tools] **Edit**: `/Users/a11401/Documents/2026/Mider/mider/tools/static_analysis/proc_runner.py`
+- `2026-02-27 10:12:15` [feat/T6-static-analysis-tools] **Edit**: `/Users/a11401/Documents/2026/Mider/mider/tools/static_analysis/proc_runner.py`
+- `2026-02-27 10:12:19` [feat/T6-static-analysis-tools] **Edit**: `/Users/a11401/Documents/2026/Mider/tests/test_tools/test_eslint_runner.py`
+- `2026-02-27 10:12:31` [feat/T6-static-analysis-tools] **Edit**: `/Users/a11401/Documents/2026/Mider/tests/test_tools/test_clang_tidy_runner.py`